backend/requirements.txt
# Backend Dependencies for Image Tagger Organizer

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
pymongo==4.6.1
motor==3.3.2

# Object Storage
minio==7.2.3

# Machine Learning - UPDATED for Python 3.12+
ultralytics==8.2.0  # Updated from 8.1.0
torch>=2.3.0  # Updated from 2.1.2
torchvision>=0.18.0  # Updated from 0.16.2
opencv-python==4.9.0.80
pillow==10.2.0
imagehash==4.3.1

# Task Queue
celery==5.3.6
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Testing (optional)
pytest==7.4.4
httpx==0.26.0